+++ src/PermissionsPolicy/PermissionsPolicyMiddleware.cs
@@ -1,5 +1,7 @@
 ﻿using Microsoft.AspNetCore.Http;
+using Microsoft.Net.Http.Headers;
 using System;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Microsoft.AspNetCore.Builder
@@ -7,27 +9,28 @@ namespace Microsoft.AspNetCore.Builder
     public static partial class AppBuilderExtensions
     {
         /// <summary>
-        /// Adds the Feature-Policy header to responses with content type text/html.
+        /// Adds the Permissions-Policy header to responses with content type text/html.
         /// </summary>
         /// <param name="app"></param>
         /// <param name="options"></param>
-        [Obsolete]
-        public static void UseFeaturePolicy(this IApplicationBuilder app, FeatureDirectiveList options)
+        public static void UsePermissionsPolicy(this IApplicationBuilder app, PermissionsPolicyDirectiveList options)
         {
-            app.UseMiddleware<FeaturePolicyMiddleware>(options);
+            app.UseMiddleware<PermissionsPolicyMiddleware>(options);
         }
 
 
-        internal sealed class FeaturePolicyMiddleware
+        internal sealed class PermissionsPolicyMiddleware
         {
-            public FeaturePolicyMiddleware(RequestDelegate next, FeatureDirectiveList options)
+            public PermissionsPolicyMiddleware(RequestDelegate next, PermissionsPolicyDirectiveList options)
             {
                 _next = next;
                 Options = options ?? throw new ArgumentNullException(nameof(options));
             }
 
+            private const string HeaderName = "Permissions-Policy";
+
             private readonly RequestDelegate _next;
-            public FeatureDirectiveList Options { get; }
+            public PermissionsPolicyDirectiveList Options { get; }
             
             public async Task Invoke(HttpContext context)
             {
@@ -35,9 +38,10 @@ namespace Microsoft.AspNetCore.Builder
                 {
                     HttpResponse response = context.Response;
 
-                    if (response.GetTypedHeaders().ContentType?.MediaType.Equals("text/html", StringComparison.OrdinalIgnoreCase) ?? false)
+                    if (response.Headers.TryGetValue(HeaderNames.ContentType, out var values) &&
+                        values.Any(v => v.StartsWith("text/html", StringComparison.OrdinalIgnoreCase)))
                     {
-                        response.Headers["Feature-Policy"] = Options.ToString();
+                        response.Headers[HeaderName] = Options.ToString();
                     }
 
                     return Task.CompletedTask;